+++ libs/game/include/@@headers.h
@@ -50,7 +50,7 @@ const int  APP_BUILD        =1;
 /******************************************************************************/
 // CPP
 /******************************************************************************/
-#include "Auto.h"
+#include "LoadAssets.h"
 #include "Main.h"
 /******************************************************************************/
 // INLINE, TEMPLATES
