+++ libs/game/src/LoadAssets.cpp
